+++ libs/core/src/state/widget/mod.rs
@@ -230,6 +230,7 @@ pub struct WidgetDebugInfo {
     pub monitor_id: Option<String>,
     pub instance_id: Option<String>,
     pub status: WidgetStatus,
+    pub webview_window_id: Option<isize>,
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, TS)]