+++ changelog.md
@@ -19,6 +19,7 @@
 - toolbar and dock autohide weird behaviour on hitboxes.
 - twm not reflecting changes inmediately on settings change.
 - dock/toolbar/twm behaviours while using fullscreen apps.
+- not showing toolbar and dock when Apps Menu open.
 
 ## [2.6.10]